rename die2 -> die
