+++ application/api/models/die.ts
